scraper/requirements.txt
playwright==1.40.0
asyncpg==0.29.0
aiohttp==3.9.1
aiofiles==23.2.1
python-dotenv==1.0.0
supabase==2.3.0
postgrest==0.13.0